Move user agent settings to launch tab
